🎨 root elements refactored
ThemeProvider | MuiThemeProvider | CssBaseline | GlobalSty
le | Provider moved from layout component wrapper to there own WrapRootElement compone
nt that wraps the site during build process. Makes life simpler :-)
